requirements.txt
pywikibot
discord.py
boto3
requests-sse
sseclient
pytest
pytest-asyncio
